push

+++ my_put_nbrex.c
@@ -4,7 +4,7 @@ void	my_putchar_base(int c, char *base)
 {
 	my_putchar(base[c]);
 }
-int	my_put_nbr_ex(unsigned long long int nb)
+int	my_put_nbr_ex(unsigned int nb)
 {
 	char	*base = "0123456789abcdef";
 	int b = my_strlen(base);
